fix(docker/tls): move TLSFieldset to react component folder
