+++ app/portainer/components/tls-fieldset/index.ts
@@ -1,21 +1,22 @@
 import angular from 'angular';
 
-import { TLSFieldset } from '@/react/portainer/environments/wizard';
+import {
+  TLSFieldset,
+  tlsConfigValidation,
+} from '@/react/components/TLSFieldset';
 import { withFormValidation } from '@/react-tools/withFormValidation';
-import { tlsConfigValidation } from '@/react/portainer/environments/wizard/EnvironmentsCreationView/WizardDocker/APITab';
 
-import { tlsFieldsetAngular } from './tlsFieldsetAngular';
-
-export const ngModule = angular
-  .module('portainer.app.components.tls-fieldset', [])
-  .component('tlsFieldset', tlsFieldsetAngular);
+export const ngModule = angular.module(
+  'portainer.app.components.tls-fieldset',
+  []
+);
 
 export const tlsFieldsetModule = ngModule.name;
 
 withFormValidation(
   ngModule,
   TLSFieldset,
-  'tlsFieldsetReact',
-  ['values', 'onChange', 'errors'],
+  'tlsFieldset',
+  [],
   tlsConfigValidation
 );
