+++ app/portainer/views/endpoints/edit/endpoint.html
@@ -72,7 +72,7 @@
   <div class="col-lg-12 col-md-12 col-xs-12">
     <rd-widget>
       <rd-widget-body>
-        <form class="form-horizontal">
+        <form class="form-horizontal" name="$ctrl.endpointForm">
           <div class="col-sm-12 form-section-title"> Configuration </div>
           <!-- name-input -->
           <div class="form-group">
@@ -220,7 +220,7 @@
               <button
                 type="button"
                 class="btn btn-primary btn-sm !ml-0"
-                ng-disabled="state.actionInProgress || !endpoint.Name || !endpoint.URL || (endpoint.TLS && ((endpoint.TLSVerify && !formValues.TLSCACert) || (endpoint.TLSClientCert && (!formValues.TLSCert || !formValues.TLSKey)))) || (formValues.tlsConfig.tls && !formValues.tlsConfig.skipVerify && (!formValues.tlsConfig.caCertFile || !formValues.tlsConfig.certFile || !formValues.tlsConfig.keyFile)) "
+                ng-disabled="state.actionInProgress || !endpoint.Name || !endpoint.URL || !$ctrl.endpointForm.$valid"
                 ng-click="updateEndpoint()"
                 button-spinner="state.actionInProgress"
               >
