+++ app/react/components/TLSFieldset/index.ts
@@ -1,0 +1,1 @@
+export { TLSFieldset, tlsConfigValidation } from './TLSFieldset';
