+++ app/react/components/TLSFieldset/types.ts
@@ -1,0 +1,7 @@
+export interface TLSConfig {
+  tls: boolean;
+  skipVerify?: boolean;
+  caCertFile?: File;
+  certFile?: File;
+  keyFile?: File;
+}
