+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/WizardDocker/APITab/APIForm.validation.tsx
@@ -1,9 +1,10 @@
 import { object, SchemaOf, string } from 'yup';
 
+import { tlsConfigValidation } from '@/react/components/TLSFieldset/TLSFieldset';
+
 import { metadataValidation } from '../../shared/MetadataFieldset/validation';
 import { useNameValidation } from '../../shared/NameField';
 
-import { tlsConfigValidation } from './TLSFieldset';
 import { FormValues } from './types';
 
 export function useValidation(): SchemaOf<FormValues> {
