+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/WizardDocker/APITab/index.ts
@@ -1,3 +1,1 @@
 export { APITab } from './APITab';
-export { TLSFieldset, tlsConfigValidation } from './TLSFieldset';
-export * from './types';
