+++ app/react/portainer/environments/wizard/EnvironmentsCreationView/WizardDocker/APITab/types.ts
@@ -1,12 +1,5 @@
+import { TLSConfig } from '@/react/components/TLSFieldset/types';
 import { EnvironmentMetadata } from '@/react/portainer/environments/environment.service/create';
-
-export interface TLSConfig {
-  tls: boolean;
-  skipVerify?: boolean;
-  caCertFile?: File;
-  certFile?: File;
-  keyFile?: File;
-}
 
 export interface FormValues {
   name: string;
