+++ app/react/portainer/environments/wizard/index.ts
@@ -1,4 +1,3 @@
 export { HomeView } from './HomeView';
 export { EnvironmentTypeSelectView } from './EnvironmentTypeSelectView';
 export { EnvironmentCreationView } from './EnvironmentsCreationView';
-export { TLSFieldset } from './EnvironmentsCreationView/WizardDocker/APITab';
